refactor: improve gemini_client_test.go organization by splitting into focused test files
- Split the large gemini_client_test.go file (1574 lines) into multiple focused test files
- Created mocks_test.go for all mock implementations used in tests
- Created model_info_test.go for model info related tests
- Created generate_content_test.go for content generation tests
- Created token_counting_test.go for token counting tests
- Created helper_methods_test.go for utility method tests
- Modified original file to be a stub
- Marked task as complete in TODO.md
